backend/requirements.txt

fastapi==0.115.5
uvicorn==0.32.0
SQLAlchemy==2.0.36
pydantic==2.9.2
python-dotenv==1.0.1
openai==1.55.3
httpx==0.27.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pinecone==8.0.0
